refactor: move computeBasicAuthHeader to its own module
