+++ README.md
@@ -28,7 +28,7 @@ This indicates that IAM token authentication is to be used. Users can pass in an
 
 ### icp4d
 This indicates that the service is an instance of ICP4D, which has its own version of token authentication. Users can pass in a `username` and `password`, or an `icp_access_token`. If a username and password is given, the SDK will manage the token for the user.
-A `url` is **required** for this type. In order to use ICP4D authentication, this option **must** be passed in.
+A `url` is **required** for this type. In order to use an SDK-managed token with ICP4D authentication, this option **must** be passed in.
 
 ## Available Modules
 ### BaseService
