+++ auth/index.ts
@@ -1,0 +1,20 @@
+/**
+ * Copyright 2019 IBM Corp. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export { IamTokenManagerV1 } from './iam-token-manager-v1';
+export { Icp4dTokenManagerV1 } from './icp4d-token-manager-v1';
+export { JwtTokenManagerV1 } from './jwt-token-manager-v1';
+export * from './utils';
